+++ mllam_data_prep/ops/derive_variable/__init__.py
@@ -1,1 +1,1 @@
-from . import physical_field, time_components
+from . import dispatcher, physical_field, time_components
